creating Trip for each ServiceFare

+++ src/main/java/com/mybus/model/ServiceFare.java
@@ -52,7 +52,7 @@ public class ServiceFare {
 	
 	@Getter
 	@Setter
-	private int journeyDuration;
+	private double journeyDuration;
 	
 	@Getter
 	@Setter
